+++ src/DotRecast.Detour.TileCache/DtTileCache.cs
@@ -54,7 +54,6 @@ namespace DotRecast.Detour.TileCache
         private readonly List<long> m_update = new List<long>();
 
         private readonly DtTileCacheBuilder builder = new DtTileCacheBuilder();
-        private readonly DtTileCacheLayerHeaderReader tileReader = new DtTileCacheLayerHeaderReader();
 
         public DtTileCache(DtTileCacheParams option, DtTileCacheStorageParams storageParams, DtNavMesh navmesh, IRcCompressor tcomp, IDtTileCacheMeshProcess tmprocs)
         {
@@ -245,7 +244,7 @@ namespace DotRecast.Detour.TileCache
             // Make sure the data is in right format.
             RcByteBuffer buf = new RcByteBuffer(data);
             buf.Order(m_storageParams.Order);
-            DtTileCacheLayerHeader header = tileReader.Read(buf, m_storageParams.Compatibility);
+            DtTileCacheLayerHeader header = DtTileCacheLayerHeaderReader.Read(buf, m_storageParams.Compatibility);
             // Make sure the location is free.
             if (GetTileAt(header.tx, header.ty, header.tlayer) != null)
             {
@@ -617,24 +616,24 @@ namespace DotRecast.Detour.TileCache
                 {
                     if (ob.type == DtTileCacheObstacleType.CYLINDER)
                     {
-                        builder.MarkCylinderArea(layer, tile.header.bmin, m_params.cs, m_params.ch, ob.pos, ob.radius, ob.height, 0);
+                        DtTileCacheBuilder.MarkCylinderArea(layer, tile.header.bmin, m_params.cs, m_params.ch, ob.pos, ob.radius, ob.height, 0);
                     }
                     else if (ob.type == DtTileCacheObstacleType.BOX)
                     {
-                        builder.MarkBoxArea(layer, tile.header.bmin, m_params.cs, m_params.ch, ob.bmin, ob.bmax, 0);
+                        DtTileCacheBuilder.MarkBoxArea(layer, tile.header.bmin, m_params.cs, m_params.ch, ob.bmin, ob.bmax, 0);
                     }
                     else if (ob.type == DtTileCacheObstacleType.ORIENTED_BOX)
                     {
-                        builder.MarkBoxArea(layer, tile.header.bmin, m_params.cs, m_params.ch, ob.center, ob.extents, ob.rotAux, 0);
+                        DtTileCacheBuilder.MarkBoxArea(layer, tile.header.bmin, m_params.cs, m_params.ch, ob.center, ob.extents, ob.rotAux, 0);
                     }
                 }
             }
 
             // Build navmesh
-            builder.BuildTileCacheRegions(layer, walkableClimbVx);
-            DtTileCacheContourSet lcset = builder.BuildTileCacheContours(layer, walkableClimbVx,
-                m_params.maxSimplificationError);
-            DtTileCachePolyMesh polyMesh = builder.BuildTileCachePolyMesh(lcset, m_navmesh.GetMaxVertsPerPoly());
+            DtTileCacheBuilder.BuildTileCacheRegions(layer, walkableClimbVx);
+            DtTileCacheContourSet lcset = DtTileCacheBuilder.BuildTileCacheContours(layer, walkableClimbVx, m_params.maxSimplificationError);
+            DtTileCachePolyMesh polyMesh = DtTileCacheBuilder.BuildTileCachePolyMesh(lcset, m_navmesh.GetMaxVertsPerPoly());
+
             // Early out if the mesh tile is empty.
             if (polyMesh.npolys == 0)
             {
@@ -678,7 +677,7 @@ namespace DotRecast.Detour.TileCache
 
         public DtTileCacheLayer DecompressTile(DtCompressedTile tile)
         {
-            DtTileCacheLayer layer = builder.DecompressTileCacheLayer(m_tcomp, tile.data, m_storageParams.Order, m_storageParams.Compatibility);
+            DtTileCacheLayer layer = DtTileCacheBuilder.DecompressTileCacheLayer(m_tcomp, tile.data, m_storageParams.Order, m_storageParams.Compatibility);
             return layer;
         }