+++ src/DotRecast.Detour.TileCache/Io/DtTileCacheLayerHeaderReader.cs
@@ -23,9 +23,9 @@ using DotRecast.Core;
 
 namespace DotRecast.Detour.TileCache.Io
 {
-    public class DtTileCacheLayerHeaderReader
+    public static class DtTileCacheLayerHeaderReader
     {
-        public DtTileCacheLayerHeader Read(RcByteBuffer data, bool cCompatibility)
+        public static DtTileCacheLayerHeader Read(RcByteBuffer data, bool cCompatibility)
         {
             DtTileCacheLayerHeader header = new DtTileCacheLayerHeader();
             header.magic = data.GetInt();
@@ -39,7 +39,7 @@ namespace DotRecast.Detour.TileCache.Io
             header.tx = data.GetInt();
             header.ty = data.GetInt();
             header.tlayer = data.GetInt();
-            
+
             header.bmin.X = data.GetFloat();
             header.bmin.Y = data.GetFloat();
             header.bmin.Z = data.GetFloat();